pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rvln"
version = "0.1.0"
description = "Neuro-symbolic LTL planner + vision-language supervision for UAV control in simulation"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    # Sim / gym
    "gym<0.27",
    "unrealcv",
    "opencv-python",
    "matplotlib",
    "numpy",
    "scipy",
    "pillow",
    "requests",
    "pynput",
    # AI / LLM stack
    "openai",
    "google-generativeai",
    "huggingface-hub",
    "jinja2",
    "networkx",
    "pandas",
    "pydantic",
    "scikit-learn",
    "sentence-transformers",
    "sympy",
    "tqdm",
    "websockets",
    "beautifulsoup4",
    "tenacity",
    "imageio",
    "imageio-ffmpeg",
]

[project.optional-dependencies]
server = [
    "torch>=2.2.0",
    "torchvision>=0.16.0",
    "transformers==4.40.1",
    "timm==0.9.10",
    "tokenizers==0.19.1",
    "flask",
    "flash-attn",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["rvln*", "gym_unrealcv*"]
